JAIKO/backend/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flask-socketio==5.3.6
flask-cors==4.0.1
flask-jwt-extended==4.6.0
psycopg2-binary>=2.9.10
python-dotenv==1.0.1
google-auth==2.29.0
requests==2.31.0
supabase==2.4.1
eventlet==0.36.1
python-dateutil==2.9.0
marshmallow==3.21.2
